added character moves
TODO - set boundaries and lines for the robot to move on. the robot should not move off the map and the robot should only be able to turn at an intersection. Otherwise the robot moves, looks, and acts according to the project description

added glutSpecialUpFunc() which is a callback function that listens for when the function keys are released. this is used to turn the robots head back from being turned when the key is released.